+++ src/matterbridge.matterjs.test.ts
@@ -196,7 +196,7 @@ describe('Matterbridge matterjs', () => {
     jest.advanceTimersByTime(15 * 60 * 1000); // Advance time by 15 minutes
     jest.useRealTimers();
 
-    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.NOTICE, expect.stringContaining(`Advertising on server node for ${matterbridge.serverNode?.id} stopped. Restart to commission.`));
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.NOTICE, expect.stringContaining(`Advertising stopped. Restart to commission again.`));
   });
 
   test('Matterbridge.destroyInstance() -bridge mode', async () => {